New Quotify App using firebase
Created a new app using firebase and react native.
The app is a quote app that allows users to sign
in and out, view their profile, and view quotes.
The app is still in development and will be
updated with more features.

Right now, the app has the following features:
- Sign in/ create a new account
- Sign out
- View profile
- View quotes
- Edit profile
- Reset password
- Sync user profile with firebase.